adding a validation strategy for mid extensions
If two diags have alredy been added to the result and a thrid one is the next
possible extension, we need to also check for a mid extension.
Usually these are not possible. The method was not able to identify these cases
though. But now it prevents mid extensions, if they are impossible