adding additional mode 'paraloge' to gtamap
This can be adapted or removed later. But for now it is an easy way of
generating the fa and .gtai files required for the second pass.
Additional regions of interest can still be added but the user will have to
manually run index-pre to extract the additional sequence or provide their
own fa sequence. After that the user also needs to manually create an index
for the desired region and add the abs path to the paraloge.csv which is
uses to extend the main index by the additional paraloge indices.