adding feature to extend main index using paralog indices
For this i added an additional param to map mode.
If provided, extends the main index with all indices listed in the paralogs.csv.
I also had to update some unrelated methods in genomeIndex to
use pointer receivers.
Otherwise we wouldn't be able to extend the main index with sub indices.
Since go is strict, it either wants all methods of a struct to be pointer receivers
or value receivers. So i had to change all to pointer receivers.