adding logic to annotate junctions
After calling extendDiagonals (which also finds the bestSplit if necessary)
I call another func called annotateSpliceSites.
For every result I check if the genome reg vec has a gap.
If so, I simply check if this gap has a known splice site and store the
information in the curr ReadMatchResult