enhanching search space in graph
before this change, a defaultExtensionRight/Left was only done if no startnode was found.

but this is problematic in some cases where an l and r padding were applied
with one of them being incorrect.
By always remapping the default extension, any padding that is incorrectly appended
is also checked one time with out appending it